server/pyproject.toml
[tool.poetry]
name = "DearMEP"
version = "1.0.0-dev.9999999999"
description = "DearMEP helps people contact their representatives."
authors = ["Tim Weber"]
license = "AGPL-3.0"
# Include the frontend, see <https://github.com/python-poetry/poetry/issues/2809>
include = ["dearmep/static_files/static/**/*"]

[tool.poetry.scripts]
dearmep = "dearmep.cli:run"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.1"
PyYAML = "^6.0"
limits = "^3.3.1"
lzip = "^1.2.0"
rich = "^13.4.2"
countryguess = "^0.3.0"
defusedxml = "^0.7.1"
httpx = ">=0.23"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
python-geoacumen = ">=2023"
starlette-exporter = "^0.15.1"
sqlmodel = "^0.0.8"
eralchemy2 = {version = "^1.3.7", optional = true}
csvkit = {version = "*", optional = true}
visidata = {version = "*", optional = true}
pillow = "^9.5"
requests = "^2.31"
backoff = "^2.2.1"
ratelimit = "^2.2.1"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
markdown-it-py = "^2.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4.8"
xlsxwriter = {version = "^3.1.5", optional = true}
canonicaljson = "^2.0.0"
phonenumbers = "^8.13.22"

[tool.poetry.group.dev.dependencies]
flake8 = "^4.0.1"
flake8-quotes = "^3.3.1"
mypy = "<1.0"
pytest = "^7.3.0"
pytest-cov = "^4.0.0"
py-mmdb-encoder = "^1.0.4"
types-PyYAML = "^6.0.11"
httpx = "^0.23.1"
types-attrs = "^19.1.0"
types-maxminddb = "^1.5.0"
types-pillow = "^9.5"
types-requests = "^2.31"

[tool.poetry.extras]
convert = ["csvkit", "visidata", "xlsxwriter"]
specs = ["eralchemy2"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
